requirements.txt
langchain
unstructured
ollama
chromadb
streamlit
pdfplumber
langchain-core
langchain-community
onnx==1.16.1
langchain-text-splitters
faiss-cpu
pdfminer.six==20221105
unstructured-inference
unstructured_pytesseract
rouge
nltk
pytesseract
scikit-learn
libtesseract-dev
tesseract-ocr
seaborn
moverscore
matplotlib
pyemd
pdf2image
pi-heif
torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu117
poppler-utils
